basic rabl template engine
This engine renders the given path, in the source path, using Rabl.
Rabl is in "hash" mode so that only raw data is returned.  Sanford
then takes the raw data and serializes it to BSON as needed.

This should be all that is needed to have a working sanford rabl
template engine.